pyproject.toml
[tool.poetry]
name = "correct-score"
version = "0.1.0"
description = ""
authors = ["Davies <daviesaraujo@hotmail.com>"]
readme = "README.md"
packages = [{include = "correct_score"}]

[tool.poetry.dependencies]
python = "^3.11"
bs4 = "^0.0.1"
mysql-connector-python = "^8.0.31"
selenium = "^4.7.2"
protobuf = "3.18.3"
python-dotenv = "^0.21.0"
pyrogram = "^2.0.93"
tgcrypto = "^1.2.5"

[tool.poetry.group.dev.dependencies]
webdriver-manager = "^3.8.5"
pip-audit = "^2.4.12"
mkdocs = "^1.4.2"
isort = "^5.11.4"
pytest = "^7.2.0"
prospector = "1.7.7"
ipykernel = "^6.19.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
